drop support terraform 0.11
quoted type constraints are deprecated since 0.12.14

Terraform 0.11 and earlier required type constraints to be given in quotes,
but that form is now deprecated and will be removed in a future version of
Terraform. To silence this warning, remove the quotes around "list" and write
list(string) instead to explicitly indicate that the list elements are
strings.

Signed-off-by: Outsider <outsideris@gmail.com>